+++ app/src/main/java/com/example/proyectoa_pmdm_t2_junzhou/OnDatosListener.java
@@ -1,5 +1,5 @@
 package com.example.proyectoa_pmdm_t2_junzhou;
 
 public interface OnDatosListener {
-    public void onDatosListener(Double lat, Double lon, Double dist);
+    public void onDatosListener(Double lat, Double lon, int dist);
 }
